backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.140
botocore==1.29.140
certifi==2021.5.30
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==3.0.2
Django==3.2.12
django-comment-migrate==0.1.7
django-cors-headers==3.10.1
django-filter==22.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-restql==0.15.3
django-simple-captcha==0.5.17
django-tenants==3.4.8
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.0.1
msgpack==1.0.5
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.0
Pillow==9.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyPEG2==2.15.2
pypinyin==0.48.0
python-dateutil==2.8.2
pytz==2021.1
redis==4.5.5
requests==2.28.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.15
user-agents==2.2.0
uvicorn==0.21.1
websockets==10.4
whitenoise==5.3.0
zope.event==4.6
zope.interface==6.0
